% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono.R
\name{chrono}
\alias{chrono}
\title{Create a timeline of a selected dataset}
\usage{
chrono(d = NA, seriated = TRUE, all_dates = FALSE, verbose = TRUE)
}
\arguments{
\item{d}{A dataframe}

\item{seriated}{If TRUE (default), re-order by minimum (earliest) date.}

\item{all_dates}{If TRUE, show all sites, including those with no date. Default: FALSE.}

\item{verbose}{If TRUE (default), verbose.}
}
\value{
A ggplot graph.
}
\description{
Read EDTF data and return a ggplot timeline
}
\examples{
df <- db_api_connect()
# Show the dataset timeline (seriated)
chrono(df$dataset_adisser17)

# Show the dataset timeline (not seriated)
chrono(df$dataset_adisser17, seriated = FALSE)

# Show all sites (including those with no date)
chrono(df$dataset_adisser17, all_dates = TRUE)

}
