% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_api_connect.R
\name{db_api_connect}
\alias{db_api_connect}
\title{Connect the DB API and return an R object (dataframe, etc.)}
\usage{
db_api_connect(
  d = NA,
  api_url = "http://157.136.252.188:3000/dataset_adisser17",
  output_format = "dataframe",
  verbose = TRUE
)
}
\arguments{
\item{d}{A hash object. If none is provided, a new one will be created.}

\item{api_url}{An URL landing to an API. Default: 'dataset_adisser17'.}

\item{output_format}{The selected output format. Default "dataframe".}

\item{verbose}{if TRUE (by default), verbose.}
}
\value{
An R object.
}
\description{
Connect the DB API and return an R object (dataframe, etc.)
}
\examples{
df <- db_api_connect()
# Show the first rows of the dataframe
head(df$dataset_adisser17)

}
