% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa.R
\name{ppa}
\alias{ppa}
\title{Classify a point pattern distribution}
\usage{
ppa(
  d = NA,
  root = "https://raw.githubusercontent.com/iramat/iRamat/master/inst/extdata/",
  img.paths = c("clustered_distribution.png", "random_distribution.png",
    "regular_distribution.png"),
  ppa_tests = c("quadrat", "ripley", "gfunction"),
  verbose = TRUE
)
}
\arguments{
\item{root}{The path to the raster dataset.}

\item{img.paths}{Name of the images.}

\item{ppa_tests}{A list of test to perform, in a vector. By default: Quadrat Test ("quadrat"), Ripley’s K ("ripley"), G-function ("gfunction"). The vector looks like: c("quadrat", "ripley", "gfunction"). R quadrat.test() function: If p < 0.05 → the pattern is likely non-random (either clustered or regular). R Kest() (Ripley) function: ... R Gest() (G-function) function: ...}

\item{verbose}{if TRUE (by default), verbose.}
}
\value{
Geostatistical test results.
}
\description{
Classify a point pattern distribution from raster (images)
}
\examples{
d <- ppa()
# Show results of Quadrat test of the clustered distribution
d$clustered_distribution.png$quadrat
# Plot the K-Ripley of the clustered distribution
plot(d$clustered_distribution.png$ripley, main = "clustered distribution")
# Plot the G-function of the regular distribution
plot(d$regular_distribution.png$gfunction, main = "regular distribution")
}
